펀딩 디테일 response 변경

+++ data/src/main/java/com/danbam/data/remote/response/FundingResponse.kt
@@ -11,7 +11,7 @@ data class FundingResponse(
     @SerializedName("description")
     val description: String,
     @SerializedName("percentage")
-    val percentage: Long,
+    val percentage: Double,
     @SerializedName("thumbnailUrl")
     val thumbnail: String,
     @SerializedName("status")
